+++ Converter/SettingsManager.cs
@@ -22,21 +22,45 @@ public partial class SettingsJsonContext : JsonSerializerContext { }
 
 public static class SettingsManager
 {
-    private static string settingsFileName = $"{Environment.CurrentDirectory}/settings.json";
-    public static Settings Settings { get; set; }
+    private static readonly string settingsFileName = Path.Combine(ExecutablePath, "settings.json");
+    private static readonly string defaultModsDirectory = Path.Combine(MyDocuments, "Paradox Interactive", "Crusader Kings III", "mod");
+    private static readonly string defaultInputJsonPath = Path.Combine(ExecutablePath, "input.json");
+    private static readonly string defaultInputGeojsonPath = Path.Combine(ExecutablePath, "input.geojson");
+    private static string MyDocuments => Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
 
-    private static string GetGameDirectoryWindows()
+    public static string ExecutablePath => Directory.GetParent(Environment.ProcessPath!)!.FullName;
+
+    public static Settings Instance { get; private set; }
+
+    private static string GetSteamLibraryFoldersPath()
     {
-        var steamRegistryKey = Environment.Is64BitOperatingSystem
-            ? "HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\Valve\\Steam"
-            : "HKEY_LOCAL_MACHINE\\SOFTWARE\\Valve\\Steam";
+        if (OperatingSystem.IsWindows())
+        {
+            var steamRegistryKey = Environment.Is64BitOperatingSystem
+                ? "HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\Valve\\Steam"
+                : "HKEY_LOCAL_MACHINE\\SOFTWARE\\Valve\\Steam";
 
-        var steamPath = Registry.GetValue(steamRegistryKey, "InstallPath", null);
-        var libraries = File.ReadAllText($"{steamPath}/steamapps/libraryfolders.vdf");
+            var steamPath = (string)Registry.GetValue(steamRegistryKey, "InstallPath", null)!;
+            return Path.Combine(steamPath, "steamapps", "libraryfolders.vdf");
+        }
+        else if (OperatingSystem.IsMacOS())
+        {
+            var steamPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), "Library", "Application Support", "Steam");
+            return Path.Combine(steamPath, "steamapps", "libraryfolders.vdf");
+        }
+        else
+        {
+            throw new Exception("Operating System not supported");
+        }
+    }
+
+    private static string GetGameDirectory()
+    {
+        var libraries = File.ReadAllText(GetSteamLibraryFoldersPath());
         var pathRegex = new Regex("\"path\"\\s*\"(.+)\"");
         var paths = pathRegex.Matches(libraries).Select(n => n.Groups[1].Value);
-
-        var ck3Directories = paths.Select(n => $"{n}/steamapps/common/Crusader Kings III/game").Where(Directory.Exists).ToArray();
+        
+        var ck3Directories = paths.Select(n => Path.Combine(n, "steamapps", "common", "Crusader Kings III", "game")).Where(Directory.Exists).ToArray();
         if (ck3Directories.Length > 1)
         {
             Debugger.Break();
@@ -50,40 +74,13 @@ public static class SettingsManager
 
         return ck3Directories[0];
     }
-    private static string GetGameDirectoryMac()
+    private static string GetTotalConversionSandboxDirectory()
     {
-        var steamPath = $"{Environment.GetFolderPath(Environment.SpecialFolder.UserProfile)}/Library/Application Support/Steam";
-        var libraries = File.ReadAllText($"{steamPath}/steamapps/libraryfolders.vdf");
+        var libraries = File.ReadAllText(GetSteamLibraryFoldersPath());
         var pathRegex = new Regex("\"path\"\\s*\"(.+)\"");
         var paths = pathRegex.Matches(libraries).Select(n => n.Groups[1].Value);
 
-        var ck3Directories = paths.Select(n => $"{n}/steamapps/common/Crusader Kings III/game").Where(Directory.Exists).ToArray();
-        if (ck3Directories.Length > 1)
-        {
-            Debugger.Break();
-            throw new Exception("Multiple game directories found.");
-        }
-        else if (ck3Directories.Length == 0)
-        {
-            Debugger.Break();
-            throw new Exception("No game directories found.");
-        }
-
-        return ck3Directories[0];
-    }
-
-    private static string GetTotalConversionSandboxDirectoryWindows()
-    {
-        var steamRegistryKey = Environment.Is64BitOperatingSystem
-            ? "HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\Valve\\Steam"
-            : "HKEY_LOCAL_MACHINE\\SOFTWARE\\Valve\\Steam";
-
-        var steamPath = Registry.GetValue(steamRegistryKey, "InstallPath", null);
-        var libraries = File.ReadAllText($"{steamPath}/steamapps/libraryfolders.vdf");
-        var pathRegex = new Regex("\"path\"\\s*\"(.+)\"");
-        var paths = pathRegex.Matches(libraries).Select(n => n.Groups[1].Value);
-
-        var ck3Directories = paths.Select(n => $"{n}/steamapps/workshop/content/1158310/2524797018").Where(Directory.Exists).ToArray();
+        var ck3Directories = paths.Select(n => Path.Combine(n, "steamapps", "workshop", "content", "1158310", "2524797018")).Where(Directory.Exists).ToArray();
         if (ck3Directories.Length > 1)
         {
             Debugger.Break();
@@ -97,28 +94,6 @@ public static class SettingsManager
 
         return ck3Directories[0];
     }
-    private static string GetTotalConversionSandboxDirectoryMac()
-    {
-        var steamPath = $"{Environment.GetFolderPath(Environment.SpecialFolder.UserProfile)}/Library/Application Support/Steam";
-        var libraries = File.ReadAllText($"{steamPath}/steamapps/libraryfolders.vdf");
-        var pathRegex = new Regex("\"path\"\\s*\"(.+)\"");
-        var paths = pathRegex.Matches(libraries).Select(n => n.Groups[1].Value);
-
-        var ck3Directories = paths.Select(n => $"{n}/steamapps/workshop/content/1158310/2524797018").Where(Directory.Exists).ToArray();
-        if (ck3Directories.Length > 1)
-        {
-            Debugger.Break();
-            throw new Exception("Multiple mod directories found.");
-        }
-        else if (ck3Directories.Length == 0)
-        {
-            Debugger.Break();
-            throw new Exception("No mod directories found.");
-        }
-
-        return ck3Directories[0];
-    }
-
 
     public static void Configure()
     {
@@ -128,6 +103,7 @@ public static class SettingsManager
     }
     public static bool TryLoad()
     {
+        string settings = "";
         try
         {
             if (!File.Exists(settingsFileName))
@@ -135,59 +111,34 @@ public static class SettingsManager
                 return false;
             }
 
-            var settings = File.ReadAllText(settingsFileName);
-            Settings = JsonSerializer.Deserialize(settings, SettingsJsonContext.Default.Settings)!;
+            settings = File.ReadAllText(settingsFileName);
+            Instance = JsonSerializer.Deserialize(settings, SettingsJsonContext.Default.Settings)!;
             return true;
         } 
         catch (Exception e)
         {
             Console.WriteLine($"Failed to load settings: {e.Message} {e.StackTrace}");
+            Console.WriteLine($"Content was: {settings}");
             return false;
         }
     }
     public static void CreateDefault()
     {
-        if (OperatingSystem.IsWindows())
+        Instance = new Settings
         {
-            Settings = new Settings
-            {
-                modsDirectory = $"{Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments)}/Paradox Interactive/Crusader Kings III/mod"
-                    .Replace(@"\\", "/").Replace(@"\", "/"),
-                ck3Directory = GetGameDirectoryWindows()
-                    .Replace(@"\\", "/").Replace(@"\", "/"),
-                totalConversionSandboxPath = GetTotalConversionSandboxDirectoryWindows()
-                    .Replace(@"\\", "/").Replace(@"\", "/"),
-                inputJsonPath = $"{ExecutablePath}/input.json"
-                    .Replace(@"\\", "/").Replace(@"\", "/"),
-                inputGeojsonPath = $"{ExecutablePath}/input.geojson"
-                    .Replace(@"\\", "/").Replace(@"\", "/"),
-            };
-        }
-        else if (OperatingSystem.IsMacOS())
-        {
-            Settings = new Settings
-            {
-                modsDirectory = $"{Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments)}/Paradox Interactive/Crusader Kings III/mod"
-                    .Replace(@"\\", "/").Replace(@"\", "/"),
-                ck3Directory = GetGameDirectoryMac()
-                    .Replace(@"\\", "/").Replace(@"\", "/"),
-                totalConversionSandboxPath = GetTotalConversionSandboxDirectoryMac()
-                    .Replace(@"\\", "/").Replace(@"\", "/"),
-                inputJsonPath = $"{ExecutablePath}/input.json"
-                    .Replace(@"\\", "/").Replace(@"\", "/"),
-                inputGeojsonPath = $"{ExecutablePath}/input.geojson"
-                    .Replace(@"\\", "/").Replace(@"\", "/"),
-            };
-        }
+            modsDirectory = defaultModsDirectory,
+            inputJsonPath = defaultInputJsonPath,
+            inputGeojsonPath = defaultInputGeojsonPath,
+            totalConversionSandboxPath = GetTotalConversionSandboxDirectory(),
+            ck3Directory = GetGameDirectory(),
+        };
 
         Save();
     }
 
     public static void Save()
     {
-        File.WriteAllText(settingsFileName, JsonSerializer.Serialize(Settings, SettingsJsonContext.Default.Settings));
+        File.WriteAllText(settingsFileName, JsonSerializer.Serialize(Instance, SettingsJsonContext.Default.Settings));
     }
-
-    public static string ExecutablePath => Directory.GetParent(Environment.ProcessPath!)!.FullName;
 }